// convex/_generated/api.d.ts
declare const fullApi: ApiFromModules<{
  http: typeof http;
  media: typeof media;
  media_internal: typeof media_internal;
  sync: typeof sync;
}>;

/**
 * A utility for referencing Convex functions in your app's public API.
 *
 * Usage:
 * ```js
 * const myFunctionReference = api.myModule.myFunction;
 * ```
 */
export declare const api: FilterApi<
  typeof fullApi,
  FunctionReference<any, "public">
>;